refactor: update import path for getVatsimToken to use cookieUtils in AirportPointEditor and NewDivision components
