+++ src/components/divisions/AirportPointEditor.jsx
@@ -10,7 +10,7 @@ import { MapContainer, TileLayer, Marker, LayersControl, useMapEvents, Popup } f
 import 'leaflet/dist/leaflet.css';
 import L from 'leaflet';
 import PropTypes from 'prop-types';
-import { getVatsimToken } from '../../utils/auth';
+import { getVatsimToken } from '../../utils/cookieUtils';
 
 
 // Add Mapbox token from environment
